+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ## UNRELEASED
 
-### Changed
+### Added
 
-- WIP
+- Dev dependency `avtocod/specs`
+- Method `Avtocod\Specifications\Specifications::getVendorDirectoryPath()`
+
+### Removed
+
+- Redundant tests that check specifications files (validation, duplicates, directory structures etc)
+- Dependency `opis/json-schema`
 
 [keepachangelog]:https://keepachangelog.com/en/1.0.0/
 [semver]:https://semver.org/spec/v2.0.0.html